requirements.txt
# ==========================================================
# CORE
# ==========================================================
pandas==2.2.2
numpy==1.26.4
loguru==0.7.2

# ==========================================================
# OCR & IMAGE PROCESSING
# ==========================================================
pillow==10.3.0
pytesseract==0.3.10
opencv-python==4.9.0.80

# ==========================================================
# ML / TRAINING / EVALUATION
# ==========================================================
scikit-learn==1.4.2
catboost==1.2.5
joblib==1.4.2
matplotlib==3.8.4
shap==0.45.1
tqdm==4.66.4

# ==========================================================
# MLOPS / MODEL TRACKING
# ==========================================================
mlflow==2.12.1
pyyaml==6.0.1

# ==========================================================
# DATA VALIDATION & TESTING
# ==========================================================
great-expectations==0.18.12
pytest==8.3.2

# ==========================================================
# DVC (Data Versioning)
# ==========================================================
dvc==3.51.0
dvc[gdrive]==3.51.0

# ==========================================================
# AIRFLOW PROVIDERS
# ==========================================================
apache-airflow-providers-common-sql==1.14.0
apache-airflow-providers-postgres==5.10.0
